using uint16 to reduce what's being tested against

+++ intro_to_property_based_tests/roman_numerals.go
@@ -2,14 +2,14 @@ package property_based
 
 import "strings"
 
-func ConvertToArabic(roman string) (total int) {
+func ConvertToArabic(roman string) (total uint16) {
 	for _, symbols := range windowedRoman(roman).Symbols() {
 		total += allRomanNumerals.ValueOf(symbols...)
 	}
 	return
 }
 
-func ConvertToRoman(arabic int) string {
+func ConvertToRoman(arabic uint16) string {
 	var result strings.Builder
 
 	for _, numeral := range allRomanNumerals {
@@ -22,13 +22,13 @@ func ConvertToRoman(arabic int) string {
 }
 
 type romanNumeral struct {
-	Value  int
+	Value  uint16
 	Symbol string
 }
 
 type romanNumerals []romanNumeral
 
-func (r romanNumerals) ValueOf(symbols ...byte) int {
+func (r romanNumerals) ValueOf(symbols ...byte) uint16 {
 	symbol := string(symbols)
 	for _, s := range r {
 		if s.Symbol == symbol {
@@ -72,7 +72,7 @@ func (w windowedRoman) Symbols() (symbols [][]byte) {
 		symbol := w[i]
 		notAtEnd := i+1 < len(w)
 
-		if notAtEnd && isSubtractive(symbol) && allRomanNumerals.Exists(symbol, w[i+1]) {
+		if notAtEnd && isSubtractive(uint(symbol)) && allRomanNumerals.Exists(symbol, w[i+1]) {
 			symbols = append(symbols, []byte{symbol, w[i+1]})
 			i++
 		} else {
@@ -82,6 +82,6 @@ func (w windowedRoman) Symbols() (symbols [][]byte) {
 	return
 }
 
-func isSubtractive(symbol uint8) bool {
+func isSubtractive(symbol uint) bool {
 	return symbol == 'I' || symbol == 'X' || symbol == 'C'
 }
